Fixed length pendulum with learning iterations
The cost function penalizes large deviations from the original trajectory, large deviations from the final trajectory point, large velocities and large accelerations. The velocities and accelerations need to be reconsidered, and deviations from the first trajectory point also need to be penalized. Something that can be rewarded is small variations in theta.